+++ src/index.html
@@ -1,14 +1,10 @@
 <!DOCTYPE html>
 <html lang="en">
+    <head>
+        <title>React Skeleton</title>
+    </head>
 
-<head>
-    <title>React Skeleton</title>
-
-</head>
-
-<body>
-    <div id="root"></div>
-    <script type="text/javascript" src="/bundle.js"></script>
-</body>
-
+    <body>
+        <div id="root"></div>
+    </body>
 </html>
