+++ src/jsx/index.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './components/common/App';
+import App from './components/App';
 
 ReactDOM.render(<App />, document.getElementById('root'));
